requirements.txt
Flask==1.1.2
openai==0.27.2
python-dotenv==1.0.0
gunicorn==20.1.0